Improve by Alexey
Sorry for the delay, I needed to take a break in order to think and take
a look from another point of view but unfortunately, it did not help so
much. Marek and Iliya, thanks for your comments. Actually, I didn't want
you to send us the definitions, references, and other information from
Wikipedia which I know and understand not worse than you but wanted some
explanation why it is better in comparison with the previous version. As
for section 2 A, basically I prefer the simpler adjusting rule from the
previous version but if everybody prefers the variant from Wikipedia I
am ok with that and updated the section in the appropriate way. As for
section 2 B, we need to decide what two-qubit gate we put on the paper
(control phase gate or the controlled rotation). An advantage of
controlled rotation is a lack of the phase coefficient and a
disadvantage is that it is very similar to the control gate from the
Trugenberger paper (less innovation). Also, the control phase gate may
have benefits at scaling to larger registers due to that in the case
architecture can be more distributed and not everything concentrated on
the ancillary qubit. For the current number of qubits in registers both
the schemes works approximately the same (Iliya, thanks for checking).
I prefer the previous version with the control phase rotation gate but
if everybody prefers the version with CZ it is ok with me. Please,
everybody, let me know what do you think?

Here is an updated version with the CZ gate.

Also, I am still thinking about the title, I like that it is short but
it looks a bit more general then it is needed in my opinion and does not
cover generalization of SOFM to the quantum case. Maybe it would be
better something like "Quantum-assisted unsupervised data clustering on
the basis of Self-Organizing Feature Map"? What do you think?

Here is an updated version of the manuscript. First of all, I returned
to the controlled phase gate version because it looks for me a bit more
original and also may have some advantages for bigger registers. Also, I
tried to reply to all your comments and made modifications in accordance
with them (my comments in the pdf comments file). Looking forward to
your comments.